requirements.txt
annotated-types==0.6.0
anyio==4.2.0
backoff==2.2.1
boilerpy3==1.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.109.0
h11==0.14.0
haystack-ai==2.0.0b5
haystack-bm25==1.0.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
lazy-imports==0.3.1
MarkupSafe==2.1.4
monotonic==1.6
more-itertools==10.2.0
networkx==3.2.1
numpy==1.26.3
ollama-haystack==0.0.3
openai==1.10.0
pandas==2.2.0
posthog==3.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentence-transformers==2.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
pgvecto-rs==0.1.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
jq==1.6.0
psycopg==3.1.18
